(fix) Pass x-forwarded-proto as https to Superset to fix redirects
Unfortunately, an Application Load Balancer, even an internal
Application Load Balancer like the one in front of Superset, with an
HTTP listener apparently overrides x-forwarded-proto with `http`, which
makes Superset think it's on HTTP rather than HTTPS, and Superset
returns redirects to HTTP URLs, which browsers block due to the recently
added CSP.

We could use a Network Load Balancer which will pass through the
x-forwarded-proto header unchanged, but since target instances are in
ECS/Fargate, they have to be added to the target group "by ip". This
then results in the fact that from the point of view of the Superset
service security group, they are connected to from the internal IP(s) of
the network load balancer. These IPs are not fixed, and there is no way
to attach a security group to a Network Load Balancer, so we have no
apparently way of allowing internal access to the Superset service from
its Network Load Balancer without allowing access from everything else
in the VPC. We could put the Network Load Balancer in its own subnet,
and use NACLs, but that's a lot more quite faffy infrastructure...

... so we make the Application Load Balancer an HTTPS listener, so it
overrides x-forwarded-proto with the exact value it had: `https`.