(feat) Run Theia on Fargte 1.4.0
Just doing this on Theia for now since very few people are using it.

-----

This will:

- Bump the amount of space in user's home directories from 4GB to
  (20GB - docker image size)

- Allow EFS mounts, so maybe could bump it even more

Note that not specifying the platform defaults to LATEST
https://boto3.amazonaws.com/v1/documentation/api/latest/reference/services/ecs.html#ECS.Client.run_task,
but LATEST actually means 1.3.0
https://docs.aws.amazon.com/AmazonECS/latest/developerguide/platform_versions.html

The traffic to ECR when starting a tool/visualisation, as per
https://aws.amazon.com/blogs/containers/aws-fargate-launches-platform-version-1-4/
now comes from the Task ENI, as opposed to the semi-magic "Fargate ENI".
So we need to

- allow ingress at the IP level to the ECR VPC endpoint from notebooks

- make the IAM policy associated with the VPC endpoint allow access that
  was previously didn't go through it.A

------

Initial tests seemed to show that volumes shared between containers in a
task are now owned by root, which means to write to them, the containers
must be started as root. AWS Support have so far given no explanation.
However, the container chmod the files to another user and then sudo
that user, so the network-facing service will not be run as root.