(feat) Ability to run tools or visualisations on Fargate 1.4.0
Not actually enabling for anything for now, since there are issues with
volume permissions and sudo. Specifically shared volumes between
containers end up owned by root, and also sudo fails, which means not
even starting the container as root and changing ownership will work. We
might be able to run as root, but will treat that as a last resort.

So, currently adding in the ability to switch between platform versions
in order to be able to switch to, and rollback from, any change without
a deployment.

For tools or visualisations on 1.4.0, this will:

- Bump the amount of space in user's home directories from 4GB to
  (20GB - docker image size)

- Eventually allow EFS mounts, so maybe could bump it even more, and
  move away from S3 for home directories

Note that not specifying the platform defaults to LATEST
https://boto3.amazonaws.com/v1/documentation/api/latest/reference/services/ecs.html#ECS.Client.run_task,
but LATEST actually means 1.3.0
https://docs.aws.amazon.com/AmazonECS/latest/developerguide/platform_versions.html

The traffic to ECR when starting a tool/visualisation, as per
https://aws.amazon.com/blogs/containers/aws-fargate-launches-platform-version-1-4/
now comes from the Task ENI, as opposed to the semi-magic "Fargate ENI".
So we need to

- allow ingress at the IP level to the ECR VPC endpoint from notebooks

- make the IAM policy associated with the VPC endpoint allow access that
  was previously didn't go through it.