(feat) s3sync is Fargate 1.4.0 compatible
s3sync is a sidecar container running alongside tools: they share a
volume to both write to and read from.

Fargate 1.3.0, seems to preserve the ownership from the containers, the
s3sync user and group. However, in 1.4.0, it seems to now be owned by
root. To write to the volume, there seems to be little alternative than
to start the container as root.

However, we do this for a short a time as possible just to change
permissions on the shared folder, and then run the s3sync proper as the
lower-privilage s3sync user.

AWS Support have been contacted several days ago, but so far have
provided no explanation or alternative. Maybe previously named docker
volumes were used, and now they're bind mounts?