Trial syncing datasets/permissions to QuickSight
Rudimentary implementation of a management command which can be used to
sync datasets and associated user permissions to AWS QuickSight so that
we can test out whether the tool might be a good fit for us. This is
purposefully a bit shonky because it's only for prototyping.

Some infrastructure changes are needed in order to allow QuickSight to
connect to the Data Workspace datasets DB - these have been/will be
created manually in AWS console for now.

Ticket: https://trello.com/c/gRDNW4YS/1064