(feat) s3sync is Fargate 1.4.0 compatible
s3sync is a sidecar container running alongside tools: they share a
volume to both write to and read from.

Fargate 1.3.0, seems to preserve the ownership from the containers, the
s3sync user and group. However, in 1.4.0, it seems to now be owned by
root. To write to the volume, there seems to be little alternative than
to start the container as root.

However, we do this to change ownership of the files in the volume and
then run the s3sync proper as the lower-privilage s3sync user.