(feat) JupyterLab Python Fargate 1.4.0 compatible
Since the volume shared between containers in Fargate 1.4.0 ends up
being owned by root, we have to start JupyterLab as root and change the
permission of the home directory, and then start up JupyterLab as a
lower privileged user.

The ability to run sudo was very recently added to Fargate 1.4.0
(without even a minor version bump).