Update source links to support different file types
Data flow can now upload .zip files, so we can't just generate .csv
file names any more (or the download will have the wrong extension,
which will confuse some users and programs).

For source links, therefore, we will grab the extension from the URL
(for the given case it'll be an S3 URI) and apply that to the filename.