Add an endpoint to display embeded Metabase visualisation
This is a proof-of-concept endpoint that displays a full page
embed of a Metabase dashboard.

There are no permissions checks here at the moment as it's not
connected to a visualisation template, but a dashboard needs to
be first enabled for embedding from Metabase, so this shouldn't
expose any sensitive data.

It's using the existing visualisation wrapper template, so can be
switched to static header and footer in the future.