Support multiple types of visualisation
GitLab is used by advanced users to make complex visualisations, but we
also want to allow less-technically-skilled users to make visualisations
as well. We are trialling a number of other tools in order to facilitate
this, including Google Data Studio, AWS QuickSight and Metabase. We want
to be able to support these kinds of dashboards in our visualisation
catalogue page as well as the GitLab variety. These visualisations tend
to resolve the links, either directly in the case of Data Studio or
indirectly (i.e. you use a resource ID to get a dynamic URL from the
service) in the case of Metabase and QuickSight.

Visualisation Links should help us support arbitrary services for
embedding visualisations.

Ticket: https://trello.com/c/29KnANvi/1125