(feat) Create a new task definition per tool launch
From the user's point of view, this is a refactor...

... but is a step towards EFS. For home directories. For each user to
have a different mount location in EFS, this has to be specified in the
task defintion. The location can't be specified in the overrides passed
to run_task [this has been confirmed with contact with AWS Support]. So
we create a task definition per launch.

So far, there doesn't seem to be an issue doing this for visualisation
previews, but extending this behaviour to tools, and releasing ahead of
actually moving to EFS, just in case there is some unexpected behaviour
that means we have to re-evaluate/plan the move to EFS.