Bump cache lock version and auto-expire
We are regularly seeing errors where we're unable to acquire a lock
quickly enough, which we believe means it has become stuck. This patch
will give the lock a TTL of 3 minutes, and it will automatically be
available for re-locking after that time, which should reduce any need
for us to bump the lock version.