Lock the "delete users" task
We've had some sentry errors about being unable to delete users because
they don't exist that is attributed to multiple instances of this task
running at the same time. By adding a lock around the inner-workings of
the job we can ensure only one instance is running the task at a time,
which should prevent these errors. Any instances that fail to get the
lock can just die silently.