Use immutable DB users for quicksight
At the moment we have a static DB user for quicksight and whenever we
grant new permissions we re-generate credentials for the same username
and push those to QuickSight. However, we didn't have an implementation
for revoking permissions. After a little time looking at adding more
thorough grant/revoke to the single mutable user, it felt easier to
instead go for immutable users that grant all the right permissions as a
one-off and then we simply discard that user and generate a new one if
we need a different set of permissions.

Our immutable quicksight users last 7 days rather than 31 for tool
users. This is for a few reasons:

1) We re-sync permissions for all QuickSight users every night, so
really they are unlikely to be used after 24 hours. 7 days gives us some
grace if this nightly job fails for some reason.
2) We can't tell when QuickSight credentials are in use, unlike with
tools - so rather than checking if the application is in use, we will
instead delete old quicksight users when their password expires. 7 days
means that there (hopefully) won't be too many old quicksight users
hanging around for too long.