Sync to QuickSight after user model creation
A User instance needs to have a valid profile in order for QuickSight to
attempt to sync permissions to QuickSight. A profile is created as a
post-save step of the User model. By moving the sync after the model
save, we can be sure that a profile exists with a valid (if unmatched)
SSO ID - so the sync step will not fail on "user.profile" does not
exist.