Centralise sentry config
Move the sentry setup code from Django settings and the proxy to a
separate file so that we can share code/config.

Group CancelledErrors together so that we only have a single event for
them in Sentry, making them less noisy and easier to handle. Same for
ConnectionResetError, which is also fairly prolific in the proxy and
pops up in various places with the same outcome.