Stop authors creating quicksight data sources
For the initial release, we want QuickSight authors to focus on creating
dashboards with Data Workspace data as opposed to arbitrary other data.
QuickSight now lets us apply custom permissions to users so that we can
lock down creating data sources, so let's apply that permission and
enforce it as part of the regular sync.