Keep Your Files modal HTML out of the DOM until it's needed
This is to make the Your Files page more reasonable when CSS is
disabled. Note that the modal HTML is there is Javascript is disabled,
since the ng-if will remove elements using Javascript, but in this case,
the whole interface won't work anyway.

There is some duplication here, but given the number of modals, the
current uncertainty regarding the page, and the fact that to address
this I suspect will have to do a lot of faffing with templates and/or
transclusion that can get a bit complicated, suspect not worth it.