feat: quicksight sync polling
Users have to create their own accounts for QuickSight, which they
do through a DIT SSO journey we can't easily hook into. This means
that we aren't able to configure their QuickSight account the moment
it's created. As a slight workaround to this, we change the 'Launch
QuickSight' button on the tools page to go to an intermediate page
on Data Workspace, where we can kick off a background job that polls
for their account to be created in QuickSight, before redirecting
them to the QuickSight SSO login flow. This background job will check
once a second for 5 minutes for the user account to be created, and
when it is, it will create their data source and apply our custom
permissions. We hypothesis that this should catch the majority of
users, as the only sign-up step when you get to QuickSight is
entering your email address, which should be done by most within
5 minutes.