fix: loading tools with None as access_point_id
The string replace fails when the access_point_id is None, which would
be the case for literally everyone when this is deployed.

The alternative would be to make the field not be nullable. However,
since everyone won't have an access_point_id, this means we would have
to remove the unique constraint from the DB. Choosing to avoid that
since users accidentally being assigned the same access point ID would
be very bad.