fix: saving of user's Django admin page
The default is a blank string, which causes unique constraint errors
(that we want to keep to mitigate risk of a user getting access to the
another's home directory). Instead, the Postgres `NULL` is desired, which
is derived from the Python `None`.

Note that trying to set on the model:

unique=True, blank=True and null=True

as suggested at
https://docs.djangoproject.com/en/dev/ref/models/fields/#null

doesn't address the issue, since the field in Django admin is a manually
added one, rather than automatically generated from the model, which is
the case where the documentation applies.