feature: use temporary views to extract tables from datacut queries
This was previously done using the psqlparse library but this had some limitations as it treats WITH clause statements as table names.

This commit connects to the database and runs some raw SQL to extract the table names.

As this method of parsing involves running database queries, the table names are saved against a new CustomDatasetQueryTable model rather than generated dynamically.