feat: allow admins to clear explorer credentials
Data Explorer DB credentials are cached for 24 hours, and they don't
automatically update as new permissions are granted/revoked. This may
lead to cases where a user has been granted access to a dataset but
can't access it in Data Explorer and they end up reaching out for
support.

This allows admins to clear those cached credentials so that the user
immediately gets a new set, with the new permissions available.