fix: hide data_explorer tables
Our live deploys previously had a tools-based version of Data Explorer
which would create all the Data Explorer tables in the user's schema.
Now that we are moving to a multi-tenant solution, we probably still
don't want users to see the old Data Explorer tables in their schema, so
let's exclude all tables with the `data_explorer` prefix that we used.