fix: more resilient database connections
It may happen that the cached explorer credentials are invalid for some
reason, and at the moment that will throw a 500. Let's catch this
condition and automatically recover from it by testing the cached
credentials are valid when we retrieve them - if they're not, then we
will generate new ones immediately.

We lock this entire process now to ensure that two concurrent requests,
or an eager refresh by the user, doesn't end up generating multiple sets
of credentials.