fix: init GOV.UK frontend before schema search
We've had an issue lately where the schema search panel doesn't fully
initialise its GOV.UK Design System accordion sections. After
investigating (with the kind help of the GDS Design System team ...)
this is an incompatibility with list.js, which was also building the
dynamic search for the schema panel at the same time. By instead
initialising GOV.UK Frontend first, and then when that's complete kick
off the search indexing, we can avoid this issue.