feat: launch dataset in data explorer
Add a link on master dataset pages to launch a basic query of that one
of that dataset's tables in Data Explorer. For now, this is behind the
same flag that controls whether or not a user can use the integrated
Data Explorer, so that people with the tools-based version don't see
this and think it doesn't work.