fix: login sso-cycle from iframes
The previous change worked fine if the user already had the
path-specific cookie from visiting a visualisation. However, for a new
user, they would get sent to SSO and then back, and then back to SSO, in
an infinite loop. This was due to the proxy initially setting a
SameSite=Lax cookie on the first visit inside the iframe which redirects
to SSO. This means that it won't be used when it comes back from SSO,
and sends the user to SSO again.

This is fixed by ensuring we have a SameSite=None cookie, restricted to
the path /__redirect_from_sso initially sent along with the redirect to
SSO. This allows the browser to use it when it comes back from SSO in
the iframe.

Debug mode is now disabled in tests, since django-csp doesn't set the
CSP for 404s in debug mode, which is one of the test cases.