refactor: remove webpack/sass processor/npm/node
Data Explorer had a lot of preprocessing steps that aren't generally
used in Data Workspace that have made the build more complicated. We
have relatively simple JS needs right now and it doesn't feel worth the
additional tech burden to maintain these systems for a small part of the
codebase, where a fairly trivial alternative exists to not require
these.

It does mean a couple of the Data Explorer pages will have a few more
network requests for smaller JS files, but at the moment this compromise
feels acceptable.