feat: celery task to clean up old play sql records
PlaygroundSQL rows are transient records meant to store SQL queries
temporarily, until a user wants to save it more permanently. As such we
expect to fill this table with a lot of rows that eventually don't get
used any more. This adds a task that runs roughly every 6 hours and will
delete rows older than a fortnight, to help stop it growing infinitely
over time.