feat: bump pgAdmin to 4.27
This is done to fix the error

> DOMException: Blocked a frame with origin [...] from accessing a cross-origin frame

that prevents pgAdmin from fully loading when started in a new tab.

Suspect it's due to
https://github.com/postgres/pgadmin4/blob/REL-4_8/web/pgadmin/browser/static/js/layout.js#L77
which raises an exception when window.opener is present, but from a different domain.

This appears fixed in recent version of pgAdmin, since the code that
throws the exception is wrapped in a try/except block
https://github.com/postgres/pgadmin4/blob/80ab5969920b1b258c4bb52020badbc67cf805ab/web/pgadmin/static/js/window.js#L14,

It was investigated to use `rel="noopener"` on the `form` element, but
this only works on `a` elements. In a future version, we may be able to
use `a` elements that have `rel="noopener"` on them, so have a more
consistent tool environment.