feat: allow sql query box to grow in height
At the moment the SQL box of the Data Explorer is limited strictly to 10
rows. Let's allow this to grow dynamically as user's type in longer
queries - up to 75 lines max - as the limited height of the box has been
raised as a pain point by some of our current users.