feature: poll sso users from activity stream and create django users
Tools access for new users can fail sometimes due to the IAM role creation that happens in the request/response cycle. The failure can either be due to rate limiting or the fact that IAM operations are eventually consistent. This commit adds a celery task that polls the activity stream for SSO users and creates a django user for them if necessary. If the user already exists it checks if they have permissions to run tools and if so, it checks whether they need to have an IAM role created. This check is determined by a new field that has been added to the profile model. The function that gets called if an IAM role needs to be created will be implemented seperately.