fix: semantic separation between debug and local
We use the `debug` Django setting to configure Django for running
locally, but it should be possible to run the app not in debug mode when
running it locally. This patch separates out the setting for running in
debug mode from a setting that confirms we're running locally, so that
various settings can be tied to the correct semantic check.