feat: log when users view visualisations
Add entries to our event log when users view (quicksight and datastudio)
visualisations so that we can understand how users are interacting with
our catalogue and which visualisations are popular/not. The eventlog is
automatically scraped by data-flow and returned to workspace as a master
dataset, so no further action needed for this.