Feature/sync activity stream sso users (#1052)
* feature: poll sso users from activity stream and create django users

Tools access for new users can fail sometimes due to the IAM role creation that happens in the request/response cycle. The failure can either be due to rate limiting or the fact that IAM operations are eventually consistent. This commit adds a celery task that runs every minute and polls the activity stream for SSO users and creates a django user for them if necessary. If the user already exists it checks if they have permissions to run tools and if so, it checks whether they need to have an IAM role created. This check is determined by a new field called tools_access_role_arn that has been added to the profile model. Another celery task has been added and this gets called in the sync_activity_stream_sso_users celery task, during the login flow when a user has the "Can start local tools" permission added in the admin screen. This commit also changes the current behaviour when a tool is spawned: Instead of calling AWS to check if the user has an IAM role, it now checks the new tools_access_role_arn field on the profile model and returns this if its set. If its not set, the function calls AWS to create the role and associated policy and then saves the arn against the profile model.