test: check no migrations are missing in CI
Sometimes developers (or at least, me) forget to run `django-admin
makemigrations` for minor model changes. Let's use our CI to check that
there aren't any pending model migrations that need to be generated so
that this is less likely to get merged into master.

I have set this up outside of the standard `make check` job because
that job runs in circleci that isn't set up for docker.