feature: choose tools cpu and memory in separate page
Currently the amount of cpu and memory when launching a tool is chosen using a dropdown list. This makes it difficult to show extra content explaining the different values. This commit replaces the dropdown with a separate page with radio buttons that gives more details about the different options and allows the user to save their preference for a particular tool by utilising a new model called UserToolSizeChoice.