fix: chromium shared memory for selenium tests
In the change from alpine to debian:buster-slim, we've gone from
chromium/chromedriver v77 to v83. It appears that a previously supported
flag, `disable-dev-shm-usage` is no longer available/respected, so we
get page/tab crashes again occassionally. We can bump the shm size in
docker-compose, which gets things passing again.