fix: move tools to separate directory
Move tools to a separate subdirectory to:

1) allow tools to more easily share requirements (e.g. python libs)
   between instances.
2) provide a slightly cleaner top-level
3) group together tools code

This change is mostly to support 1, with 2+3 perhaps small side
benefits. For 1, it means we can build the docker images for the tools
from the `tools/` directory, sharing a single requirements.in/txt file
for Python requirements so that we can have stronger guarantees that all
tools will expose the same set+versions of libraries.