fix: remove celery pidfile if it persists
Sometimes celery beat seems to leave behind a pidfile even after the
original process has shut down. When we then go to start up a new server
via `start.sh`, the `parallel` call starts up another celery beat
process and monitors it. It very quickly dies, due to checking for the
pidfile. When celery beat dies, `parallel` also fails, as it's set to
fail when any of its child processes terminate (success or fail), which
means the new server is shut down and the tests fail. Checking for and
clearing celery beat's pidfile manually is very hacky, but does appear
to work in this case.