fix: saving files in pgAdmin, and exit after start
The pgAdmin user has to have the same ID as the user in s3sync, 4536,
for the syncing to work properly (and the group ID is just for
consistency)

Before this, if the pgAdmin users has a different ID to that in s3sync,
there is a race condition between the s3sync container, and pgAdmin. If
pgAdmin starts first, chowns the shared directory, then s3sync will exit
shortly after launch since it won't have enough permissions to create
files, and since S3 sync is essential, pgAdmin will also then exit. If
s3sync starts first, pgAdmin will start and stay running fine (albeit
without permission to save files).

The change to make the shared directory /home/pgadmin is not vital, but it
means that the correct directory is chosen by the UI in pgAdmin by
default.