fix: create uniuque query log index in raw sql
We're unable to use a unique index on a text field in postgres as there
is an upper char limit. To get around this we can create an md5 hash of
the text field, unfortunately django does not support this so it needs
to be done manually in the migration using raw sql.