fix: provide friendlier error messages for endpoint
The datasets API endpoint can fail for a number of reasons:

1) No primary key on the table
2) The "table" is actually a view, and so doesn't have primary keys.

The latter example comes up because we're starting to deprecate some
tables that are named incorrectly, and we're changing them over to views
on the correctly-named tables.

We might want to make this endpoint work without primary keys at some
point, but this is simply so that we can more easily recognise the
actual problem in sentry for now, rather than simply getting an "invalid
sql" error.