feat: download explorer results from temp table
Adds support for downloading results from the new temporary query
results table used by the async query flow.

We liberally duplicate code here so that we:

1) don't touch the existing code path at all, reducing the risk of
   breaking it
2) can quite easily clean up the old code when we've switched it off, by
   just deleting it and renaming the new code as appropriate.

If the user tries to download the query a long time after running it
(currently the day after), the table may have been cleaned up. We try to
detect this case and provide a semi-friendly error message to the user
indicating they should simply re-run the query and try again.