feat: add "new" design system form class
Add a GOV.UK Design System model form class that we can use to render
the "request data" flow.

There is an existing "GOVUKDesignSystemModelForm" that was only ever
made to work with very basic cases, and as such doesn't handle radios
and validation very cleanly. This is only really used in one place
(visualisation catalogue items), and doesn't allow for
cleanly/maintainably rendering forms through templates.

I start again here as we need more features and have the opportunity to
clean the slate a little bit.