fix: allow configuration of ALB dereg delay
At the moment we allow 300 seconds in all environments for ALB targets
to be deregistered. This allows in-flight requests to complete safely
before the targets are killed, but also increases deployment time
significantly as we have to wait for both dev and staging to have this
300s dereg delay each. It would be nice to have a lower (or nil)
deregistration delay in these environments so that deployments from
build to prod can be faster. There are often few/no requests in
dev+staging, and even if they are, it's probably OK for AWS to kill them
after a much shorter delay as they're only for testing.