fix: rationalise header levels
We had a lot of level 2 headings that didn't actually reflect the
structure of the page. Specifically, under the 'filters' h2 heading we
now have h3s instead of h2s for each kind of filter.

We also change the 'x results' text into an h2 for the search results,
where each result's name is an h3.

It also changes the topic tags for a search result into an unordered
list, which should be a more accurate/useful representation e.g. for
screenreaders.