fix: prevent DOS from processing long lines
In theory, users could have uploaded CSV files with very long first line
and that could drain significant memory on the server, potentially
bringing it down. So let's only process the first 100KiB of the file and
trust that this should be sufficient to read at least the first 10 lines
or so.