Move to ECR
Almost everything that is currently used moved to ECR.

An exception is JupyterLab R: we agreed to remove it a while back, so
doing it now (usage is virtually non-existant). I have hidden it from
production.

Another exception is the mirrors sync, which can be done in less of a
rush

Although remotedeskop isn't being used, maintaining that just in case so
we can run desktop applications if we need to.

Some of the unused tools are partially modified: s3sync and metrics
images have been changed to get the terraform to work, but the their
images are not in ECR. Leaving tidy up to later.