feat: add column info to master dataset page
We want to surface information about the colums in master dataset tables
to users as we think this will give them more confidence about what the
data is. We show up to 12 columns on the master dataset page. After
that, if it has more than 12 columns, we direct them to a dedicated page
for the source table where they can see all of the column info.