cicd: cache build layers for faster deployments
Now that we've migrated to ECR, we can cache the layers of
data-workspace images in order to speed up our deployment pipeline. We
also add some flags to Kaniko so that it only builds our target stage
from the Dockerfile and ignores unused stages (so no longer needs to
build the test stage, which pulls in node and some other dependencies).

In tests this has taken builds from 7-10 minutes to 1-2 minutes.