fix: use error summary for validation errors
Disable HTML5 validation as per GOV.UK Design System recommendation and
disable an error summary at the top of the page linking to the form
field being validated.

We have to do some slight hacks to fill the dynamically-generated label
and hint text for this page by reaching into the widget's custom context
and manually overwriting that as well. This is some small tech debt from
the Design System form/fields implementation to work with Django's form
system.