fix: move aria-label to nav element
GOV.UK Design System currently has the aria-label for the header
navigation on a `<ul>` element within a naked `<nav>` element. This came
up in discussion recently on x-gov slack and they seem to have
recognised this isn't ideal, and are planning to move it onto the nav
element for a future release of the Design System. So let's also do the
same.