fix: include h2 for screenreaders on tools page
At the moment, we have an h1 for 'Tools' page followed by an 'h3' for
each individual tool, then followed by an 'h2' for access requirements.
Access requirements feels like a separate thing from individual tools,
so to address this gap rather than raising each tool to an h2 we'll add
a visually-hidden h2 above the tool instances that provides a semantic
grouping for them.