refactor: use fewer NACL rules
This opens up the public_whitelisted_ingress subnet to more ports, but
the only thing in that subnet is the AWS load balancer, which only has
listeners on 22 and 443, so realistically no new ports are opened up on
any bit of infrastructure.

This is to work-around the limit on the number of rules in an NACL, we
can add more IPs to the allow-list. We can't do this indefinitely, but
for now, it's the quickest/easiest thing to be able to add a few more
IPs to the list.