feat: stop using `slim` buster variant in tools
We're currently using `buster-slim` as the base image for our
docker-based tools (e.g. theia, jupyterlab). After a chat with the devs
we don't really have a good reason for this. Our images are already
pretty gargantuan - around 4GB, so the difference between `buster` and
`buster-slim` is insignificant in comparison.

The `-slim` variant specifically excludes man pages from the build, so
in some tools we've had to hack this exclusion to re-introduce them, and
after doing that we're still missing some manpages for common tools,
e.g. grep. So since size isn't a concern, let's go back to using the
"full-fat" `buster` base image. Then we can remove our hacks and benefit
from a more complete man-db with no extra work.