feat: add related data for master datasets
Adds a "related data" section to master catalogue pages, showing links
to any data cuts that make use of tables in the master dataset. On the
dataset detail page, this is limited to 4 entries, after which it shows
a "display all" link that leads to a separate page.

This new design is also applied to the existing datacut related data
section.