feature: use simple_query_string for finder search queries
The current "match_phrase" query type means that terms must appear in the same order as they do in the document for a match to occur. Using the "match" query type will get around this but it doesn't seem to support searching for an exact phrase which a user may want to do.

Using "simple_query_string" appears to support both, meaning that searching for terms in any order will still return the correct results, i.e "United Kingdom" and "Kingdom United" will return a similar number of results. However, if a user wants to search for an exact phrase they can still do so by wrapping the phrase in double quotes.